task 6 done
