+++ get_right_func.c
