+++ main.c
@@ -14,16 +14,16 @@ int main(void)
 	int len;
 	int len2;
 	unsigned int ui;
-	long int l = 2147483647;
-
+ 	long int l = 2147483647;
+   
 	long int l2 = UINT_MAX + 1024;
 	long int res = (long int)INT_MAX * 2;
-
+	
 	long int l3 = UINT_MAX + 1024;
 
 	/* Task 4 Variabls */
 	char str4[1020] = "your_initial_string_value_here\n";
-	char s[1020] = {0};
+	char s[1020];
 	char tmp[147] = "your_tmp_string_value_here\n";
 	char tmp2[97] = "your_tmp2_string_value_here\n";
 
@@ -32,41 +32,50 @@ int main(void)
 	ui = (unsigned int)INT_MAX + 1024;
 	addr = (void *)0x7ffe637541f0;
 
+
 	_printf("Length:[%d, %i]\n", len, len);
 	printf("Length:[%d, %i]\n", len2, len2);
 	_printf("Negative:[%d]\n", -762534);
 	printf("Negative:[%d]\n", -762534);
 
+	
 	_printf("Unsigned:[%u]\n", ui);
 	printf("Unsigned:[%u]\n", ui);
 
 	_printf("Unsigned octal:[%o]\n", -10);
 	printf("Unsigned octal:[%o]\n", -10);
 
-	_printf("Unsigned binary:[%b]\n", 15);
+	_printf("Unsigned binary:[%b]\n", -15);
 	printf("Unsigned binary:[%b]\n", -15);
 
 	_printf("Unsigned hexadecimal:[%x, %X]\n", ui, ui);
 	printf("Unsigned hexadecimal:[%x, %X]\n", ui, ui);
+
 
 	_printf("Character:[%c]\n", 'H');
 	printf("Character:[%c]\n", 'H');
 	_printf("String:[%s]\n", "I am a string !");
 	printf("String:[%s]\n", "I am a string !");
 
+
 	_printf("Address:[%p]\n", addr);
 	printf("Address:[%p]\n", addr);
+
 
 	len = _printf("Percent:[%%]\n");
 	len2 = printf("Percent:[%%]\n");
 
+
 	_printf("Len:[%d]\n", len);
 	printf("Len:[%d]\n", len2);
 
+	
 	_printf("Unknown:[%r]\n");
 	printf("Unknown:[%r]\n");
 
+
 	printf("\n<==== Task 0 ====>\n");
+
 
 	printf("Let's print a simple sentence.\n");
 	_printf("Let's print a simple sentence.\n");
@@ -101,7 +110,8 @@ int main(void)
 	_printf("%s%c%c%c%s%%%s%c", "Loading ", '.', '.', '.', " 99", " Please wait", '\n');
 
 	printf("css%ccs%scscscs\n", 'T', "Test");
-	_printf("css%ccs%scscscs\n", 'T', "Test");
+	 _printf("css%ccs%scscscs\n", 'T', "Test");
+
 
 	printf(str);
 	_printf(str);
@@ -109,25 +119,28 @@ int main(void)
 	printf("man gcc:\n%s", str);
 	_printf("man gcc:\n%s", str);
 
-	printf("===printing NULL bellow=====\n");
 
+
+    printf("===printing NULL bellow=====\n");
+	
 	printf(NULL);
 	_printf(NULL);
-	printf("===printing  bellow=====\n");
+ 	printf("===printing  bellow=====\n");
 	printf("%c\n", '\0');
 	_printf("%c\n", '\0');
 
-	printf("===printing pertenteg bellow=====\n");
+ 	printf("===printing pertenteg bellow=====\n");
 	printf("%");
 	_printf("%");
 
-	printf("%!\n");
-	_printf("%!\n");
+	 printf("%!\n");
+	_printf("%!\n"); 
 
 	printf("%K\n");
 	_printf("%K\n");
 
 	printf("\n<==== Task 1 ====>\n");
+
 
 	_printf("%d\n", 1024);
 	printf("%d", 1024);
@@ -264,6 +277,9 @@ int main(void)
 	_printf("x = %x, -x = %x, X = %X, -X = %X\n", 255, -255, 255, -255);
 	printf("x = %x, -x = %x, X = %X, -X = %X\n", 255, -255, 255, -255);
 
+
+
+
 	printf("\n<==== Task 3 ====>\n");
 
 	_printf("%u\n", 1024);
@@ -368,6 +384,9 @@ int main(void)
 	_printf("uuoxxX%xuoXo\n", 1024);
 	printf("uuoxxX%xuoXo\n", 1024);
 
+
+
+
 	printf("\n<========= Task 4 ============>\n");
 
 	_printf(str4);
@@ -379,11 +398,11 @@ int main(void)
 	_printf(str4, 123456789);
 	printf(str4, 123456789);
 
-	printf("\n==================\n");
+		printf("\n==================\n");
 	printf("%s %s %s %c %c %s %s %c %s %d\n", str4, "you", "normally does preprocessing, compilation, assembly", '"', '"', "For example,", tmp, 'o', tmp2, 8);
-	printf("\n==================\n");
+		printf("\n==================\n");
 	_printf("%s %s %s %c %c %s %s %c %s %d\n", str4, "you", "normally does preprocessing, compilation, assembly", '"', '"', "For example,", tmp, 'o', tmp2, 8);
-	printf("\n==================\n");
+		printf("\n==================\n");
 
 	printf("%s %s %s %c %c %s %s %c %s %d\n", str4, "you", "normally does preprocessing, compilation, assembly", '"', '"', "For example,", tmp, 'o', tmp2, 123456789);
 	_printf("%s %s %s %c %c %s %s %c %s %d\n", str4, "you", "normally does preprocessing, compilation, assembly", '"', '"', "For example,", tmp, 'o', tmp2, 123456789);
@@ -415,261 +434,35 @@ int main(void)
 	printf("\n================== TASK 6 ===========================\n\n");
 
 	_printf("%p\n", (void *)0x7fff5100b608);
-	_printf("%p", NULL);
-	_printf("\nCan you print an address?\n%p\nNice!\n", (void *)0x7fff5100b6f8);
-	_printf("\nCan you print several addresses?\n%p,%p,%p,%p\nNice!\n", (void *)0x7fff5100b6f8, (void *)0x7faf51f0f608, (void *)0x6ff42510b6f8, (void *)0x7fff510236f8);
+	printf("%p\n", (void *)0x7fff5100b608);
+
+	_printf("%p\n", NULL);
+	printf("%p\n", NULL);
+
+	_printf("Can you print an address?\n%p\nNice!\n", (void *)0x7fff5100b6f8);
+	printf("Can you print an address?\n%p\nNice!\n", (void *)0x7fff5100b6f8);
+
+	_printf("Can you print several addresses?\n%p,%p,%p,%p\nNice!\n", (void *)0x7fff5100b6f8, (void *)0x7faf51f0f608, (void *)0x6ff42510b6f8, (void *)0x7fff510236f8);
+	printf("Can you print several addresses?\n%p,%p,%p,%p\nNice!\n", (void *)0x7fff5100b6f8, (void *)0x7faf51f0f608, (void *)0x6ff42510b6f8, (void *)0x7fff510236f8);
+
 	_printf("");
-	_printf("\nCan you print an address?\n%p\nNice!\n", (void *)-1);
+	printf("");
+
+	_printf("Can you print an address?\n%p\nNice!\n", (void *)-1);
+	printf("Can you print an address?\n%p\nNice!\n", (void *)-1);
+
 	_printf("%pppp\n", (void *)0x7fff5100b6f8);
-
-	printf("\n================== TASK 7 ===========================\n\n");
-
-	printf("------------------------------Test 1----------------------------\n");
-	_printf("%+d", 1024);
-	printf("\n");
-	printf("%+d", 1024);
-	printf("\n");
-
-	printf("------------------------------Test 2----------------------------\n");
-	_printf("%+d", -1024);
-	printf("\n");
-	printf("%+d", -1024);
-	printf("\n");
-
-	printf("------------------------------Test 3--------------------------------\n");
-	_printf("%+d", 0);
-	printf("\n");
-	printf("%+d", 0);
-	printf("\n");
-
-	printf("------------------------------Test 4---------------------------\n");
-	_printf("%+d", INT_MAX);
-	printf("\n");
-	printf("%+d", INT_MAX);
-	printf("\n");
-
-	printf("------------------------------Test 5----------------------------\n");
-	_printf("%+d", INT_MIN);
-	printf("\n");
-	printf("%+d", INT_MIN);
-	printf("\n");
-
-	printf("------------------------------Test 6-----------------------------\n");
-	_printf("%+d", l);
-	printf("\n");
-	printf("%+d", l);
-	printf("\n");
-
-	printf("------------------------------Test 7-----------------------------\n");
-	_printf("%+d", l);
-	printf("\n");
-	printf("%+d", l);
-	printf("\n");
-
-	printf("------------------------------Test 8-----------------------------\n");
-	_printf("There is %+d bytes in %+d KB\n", 1024, 1);
-	printf("\n");
-	printf("There is %+d bytes in %+d KB\n", 1024, 1);
-	printf("\n");
-
-	printf("------------------------------Test 9-----------------------------\n");
-	_printf("%+d - %+d = %+d\n", 1024, 2048, -1024);
-	printf("\n");
-	printf("%+d - %+d = %+d\n", 1024, 2048, -1024);
-	printf("\n");
-
-	printf("------------------------------Test 10----------------------------\n");
-	_printf("%+d + %+d = %+d\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-	printf("%+d + %+d = %+d\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-
-	printf("------------------------------Test 11----------------------------\n");
-	_printf("%+i", 1024);
-	printf("\n");
-	printf("%+i", 1024);
-	printf("\n");
-
-	printf("------------------------------Test 12----------------------------\n");
-	_printf("%+i", -1024);
-	printf("\n");
-	printf("%+i", -1024);
-	printf("\n");
-
-	printf("------------------------------Test 13----------------------------\n");
-	_printf("%+i", 0);
-	printf("\n");
-	printf("%+i", 0);
-	printf("\n");
-
-	printf("------------------------------Test 14----------------------------\n");
-	_printf("%+i", INT_MAX);
-	printf("\n");
-	printf("%+i", INT_MAX);
-	printf("\n");
-
-	printf("------------------------------Test 15----------------------------\n");
-	_printf("%+i", INT_MIN);
-	printf("\n");
-	printf("%+i", INT_MIN);
-	printf("\n");
-
-	printf("------------------------------Test 16---------------------------\n");
-	_printf("%+i", l);
-	printf("\n");
-	printf("%+i", l);
-	printf("\n");
-
-	printf("------------------------------Test 17---------------------------\n");
-	_printf("%+i", l);
-	printf("\n");
-	printf("%+i", l);
-	printf("\n");
-
-	printf("------------------------------Test 18---------------------------\n");
-	_printf("There is %+i bytes in %+i KB\n", 1024, 1);
-	printf("\n");
-	printf("There is %+i bytes in %+i KB\n", 1024, 1);
-	printf("\n");
-
-	printf("------------------------------Test 19---------------------------\n");
-	_printf("%+i - %+i = %+i\n", 1024, 2048, -1024);
-	printf("\n");
-	printf("%+i - %+i = %+i\n", 1024, 2048, -1024);
-	printf("\n");
-
-	printf("------------------------------Test 20---------------------------\n");
-	_printf("%+i + %+i = %+i\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-	printf("%+i + %+i = %+i\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-
-	printf("------------------------------Test 21-------------------------\n");
-	_printf("% d", 1024);
-	printf("\n");
-	printf("% d", 1024);
-	printf("\n");
-
-	printf("------------------------------Test 22-------------------------\n");
-	_printf("% d", -1024);
-	printf("\n");
-	printf("% d", -1024);
-	printf("\n");
-
-	printf("------------------------------Test 23-------------------------\n");
-	_printf("% d", 0);
-	printf("\n");
-	printf("% d", 0);
-	printf("\n");
-
-	printf("------------------------------Test 24-------------------------\n");
-	_printf("% d", INT_MAX);
-	printf("\n");
-	printf("% d", INT_MAX);
-	printf("\n");
-
-	printf("------------------------------Test 25-------------------------\n");
-	_printf("% d", INT_MIN);
-	printf("\n");
-	printf("% d", INT_MIN);
-	printf("\n");
-
-	printf("------------------------------Test 26-------------------------\n");
-	_printf("% d", l);
-	printf("\n");
-	printf("% d", l);
-	printf("\n");
-
-	printf("------------------------------Test 27-------------------------\n");
-	_printf("% d", l);
-	printf("\n");
-	printf("% d", l);
-	printf("\n");
-
-	printf("------------------------------Test 28-------------------------\n");
-	_printf("There is % d bytes in % d KB\n", 1024, 1);
-	printf("\n");
-	printf("There is % d bytes in % d KB\n", 1024, 1);
-	printf("\n");
-
-	printf("------------------------------Test 29-------------------------\n");
-	_printf("% d - % d = % d\n", 1024, 2048, -1024);
-	printf("\n");
-	printf("% d - % d = % d\n", 1024, 2048, -1024);
-	printf("\n");
-
-	printf("------------------------------Test 30-------------------------\n");
-	_printf("% d + % d = % d\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-	printf("% d + % d = % d\n", INT_MIN, INT_MAX, (INT_MIN + INT_MAX));
-	printf("\n");
-
-	printf("------------------------------Test 31------------------------\n");
-	_printf("% i", 1024);
-	printf("\n");
-	printf("% i", 1024);
-	printf("\n");
-
-	printf("------------------------------Test 32-------------------\n");
-	_printf("% i", -1024);
-	printf("\n");
-	printf("% i", -1024);
-	printf("\n");
-
-	printf("------------------------------Test 33-------------------\n");
-	_printf("% i", 0);
-	printf("\n");
-	printf("% i", 0);
-	printf("\n");
-
-	printf("------------------------------Test 34-------------------\n");
-	_printf("% i", INT_MAX);
-	printf("\n");
-	printf("% i", INT_MAX);
-	printf("\n");
-
-	printf("------------------------------Test 35-------------------\n");
-	_printf("% i", INT_MIN);
-	printf("\n");
-	printf("% i", INT_MIN);
-	printf("\n");
-
-	printf("------------------------------Test 36-------------------\n");
-	_printf("% i", l);
-	printf("\n");
-	printf("% i", l);
-	printf("\n");
-
-	printf("\n-------------------test case 1--------------------\n");
-	_printf("%#o", 1024);
-
-	printf("\n-------------------test case 2--------------------\n");
-	_printf("%#o", 0);
-
-	printf("\n-------------------test case 3--------------------\n");
-	_printf("%#o", UINT_MAX);
-
-	printf("\n-------------------test case 4--------------------\n");
-	_printf("%#x", 1024);
-
-	printf("\n-------------------test case 5--------------------\n");
-	_printf("%#x", 0);
-
-	printf("\n-------------------test case 6--------------------\n");
-	_printf("%#x", UINT_MAX);
-
-	printf("\n-------------------test case 7--------------------\n");
-	_printf("%#X", 1024);
-
-	printf("\n-------------------test case 8--------------------\n");
-	_printf("%#X", 0);
-
-	printf("\n-------------------test case 9--------------------\n");
-	_printf("%#X", UINT_MAX);
-
-	printf("\n-------------------test case 10--------------------\n");
-	_printf("% ");
-	_printf("% % % % ");
+	printf("%pppp\n", (void *)0x7fff5100b6f8);
 
 	return (0);
 }
+
+
+
+
+
+
+
+
+
+
