+++ print_spec_str.c
@@ -1,24 +1,11 @@
 #include "main.h"
 
 /**
- * _putchar - writes the character c to stdout
- * @c: The character to print
- *
- * Return: On success 1.
- * On error, -1 is returned, and errno is set appropriately.
- */
-int _putchar(char c)
-{
-	return (write(1, &c, 1));
-}
-
-/**
  * printf_s - takes an array of char and print it using the buffer
  * 1024 size
  * @s: va_list containing the argument to be printed.
  * Return: the lenght of the printfed string
  */
-
 int printf_s(va_list s)
 {
 	int counter = 0;
@@ -63,7 +50,6 @@ int printf_c(va_list c)
  *
  * Return: The total length of printed characters.
  */
-
 int printf_S(va_list S)
 {
 	char *bigS = va_arg(S, char *);
